+++ src/main/java/codesquad/handler/ConnectionHandler.java
@@ -1,6 +1,5 @@
 package codesquad.handler;
 
-import codesquad.http.HttpHandler;
 import codesquad.http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
